// src/app/Models/filters/off-plan-filter.ts
export class off_plan_filter {
  'country_id': number;
  'city_id': number;
  'min_area': number;
  'max_area': number;
  'bathrooms': number;
  'balconies': number;
  'amenity_ids': number[];
  'min_price': number;
  'max_price': number;
  'min_first_pay': number;
  'max_first_pay': number;
  'delivery_date': number;
}